Remove "magic and abstraction" unit
In order to avoid confusion about classes I haven't run that much (due
to time constraints), I'm removing them from the official course. This
is a bit sad, because I really like two of them (bike repair and magic
vs. tech), but I think it's necessary for the understandabilty of the
class.